smartconnect_api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0
